(#19955) Collapse Hiera and DataBindings
The only instance of Hiera for a terminus is in the DataBindings. To
simplify the system and make it so that the Hiera code can reasonably be
expected to raise DataBindings errors, this collapses the classes
together.
