+++ spec/fixtures/unit/indirector/data_binding/hiera/global.yaml
