+++ spec/fixtures/unit/indirector/data_binding/hiera/invalid.yaml
